refactoring env as module
